Replace map mode texture array with province palette system
- Province palette: 256 x (maxProvinces/256 * numModes) instead of full-res texture array
- Memory: ~6.4MB vs 6.24GB for 15000x6500 maps with 16 modes
- GradientMapMode: CPU gradient calculation per province, GPU palette lookup per pixel
- GPU adjacency detection: Added compute shader and GPUProvinceNeighborDetector
- Dynamic texture sizing: MapTextureManager reads dimensions from image headers
- Province capacity: Increased to 100k with warning for unsupported sizes
- Save system: Fixed currentTick (ulong), added CRC32 checksum validation
- Template utils: Added province generation script from basemap

+++ Scripts/Core/SaveLoad/SaveGameData.cs
@@ -50,7 +50,7 @@ namespace Core.SaveLoad
         /// <summary>
         /// Current game tick when saved
         /// </summary>
-        public int currentTick;
+        public ulong currentTick;
 
         /// <summary>
         /// Current game speed when saved
